backend/requirements.txt
fastapi
uvicorn
minio
ultralytics
pandas 
python-multipart
pillow
